Crate pallet_evm_precompile_relay_encoder

Precompile to encode relay staking calls via the EVM

Structs

RelayEncoderPrecompile

A precompile to provide relay stake calls encoding through evm

RewardDestinationWrapper

Enums

RelayEncoderPrecompileCall

Constants

ARRAY_LIMIT

REWARD_DESTINATION_SIZE_LIMIT

Functions

u256_to_relay_amount